requirement.txt
pymoo
cobra
bs4
requests 
streamlit 
pyyaml
pandas
ipykernel